A function that returns the number of elements in a linked list

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -10,7 +10,7 @@
  */
 size_t list_len(const list_t *h)
 {
-	int counter = 0; /*variable to iterate in the list if elements*/
+	size_t counter = 0; /*variable to iterate in the list if elements*/
 
 	while (h)
 	{
